working TaskModal
