+++ client/components/TasksInProgress.js
@@ -6,6 +6,7 @@ import Button from "react-bootstrap/Button";
 import { Form, FloatingLabel, Row, Col, FormGroup } from "react-bootstrap";
 import { getTasksByUser, updateTask } from "../redux/taskReducer";
 import Modal from "react-bootstrap/Modal";
+import TaskModal from "./TaskModal";
 // ^ to link to a specific trip in trip dashboard
 
 const TasksInProgress = (props) => {
@@ -13,15 +14,6 @@ const TasksInProgress = (props) => {
   useEffect(() => {
     dispatch(getTasksByUser());
   }, []);
-
-  const [show, setShow] = useState(false);
-  const handleClose = (event) => {
-    setShow(false);
-    // dispatch(updateTask({
-
-    // }))
-  };
-  const handleShow = () => setShow(true);
 
   const tasks = props.tasks.allItineraryTasks || [];
   let inProgressTasks = tasks.filter((task) => task.status === "in progress");
@@ -46,102 +38,7 @@ const TasksInProgress = (props) => {
                   <Card.Text>
                     Provider Name: {singleTask.provider_name}
                   </Card.Text>
-                  <Button variant="primary" onClick={handleShow}>
-                    Edit Task
-                  </Button>
-
-                  <Modal show={show} onHide={handleClose}>
-                    <Modal.Header closeButton>
-                      <Modal.Title>{singleTask.type}</Modal.Title>
-                    </Modal.Header>
-                    <Modal.Body>
-                      <Form onSubmit={handleClose}>
-                        <FloatingLabel
-                          controlId="floatingSelect"
-                          label="Select Type of Task"
-                        >
-                          <Form.Select aria-label="Floating label select example">
-                            <option> {singleTask.type}</option>
-                            <option value="Lodging">Lodging</option>
-                            <option value="Dining">Dining</option>
-                            <option value="Transportation">
-                              Transportation
-                            </option>
-                            <option value="Activity">Activity</option>
-                          </Form.Select>
-                        </FloatingLabel>
-
-                        <Row className="g-2">
-                          <Col md>
-                            <FloatingLabel
-                              controlId="floatingInputGrid"
-                              label="Start Date"
-                            >
-                              <Form.Control
-                                type="text"
-                                placeholder="Start Date"
-                              />
-                            </FloatingLabel>
-                          </Col>
-                          <Col md>
-                            <FloatingLabel
-                              controlId="floatingSelectGrid"
-                              label="End Date"
-                            >
-                              <Form.Select aria-label="Floating label select example">
-                                <option>End Date</option>
-                              </Form.Select>
-                            </FloatingLabel>
-                          </Col>
-                          <Col md>
-                            <FloatingLabel
-                              controlId="floatingInputGrid"
-                              label="Start Time"
-                            >
-                              <Form.Control
-                                type="text"
-                                placeholder="Start Date"
-                              />
-                            </FloatingLabel>
-                          </Col>
-                          <Col md>
-                            <FloatingLabel
-                              controlId="floatingSelectGrid"
-                              label="End Time"
-                            >
-                              <Form.Select aria-label="Floating label select example">
-                                <option>End Date</option>
-                              </Form.Select>
-                            </FloatingLabel>
-                          </Col>
-                        </Row>
-
-                        <Form.Group className="mb-3" controlId="taskForm">
-                          <Form.Label>Email address</Form.Label>
-                          <Form.Control
-                            type="email"
-                            placeholder="name@example.com"
-                            autoFocus
-                          />
-                        </Form.Group>
-                        <Form.Group
-                          className="mb-3"
-                          controlId="exampleForm.ControlTextarea1"
-                        >
-                          <Form.Label>Example textarea</Form.Label>
-                          <Form.Control as="textarea" rows={3} />
-                        </Form.Group>
-                      </Form>
-                    </Modal.Body>
-                    <Modal.Footer>
-                      <Button variant="secondary" onClick={handleClose}>
-                        Close
-                      </Button>
-                      <Button variant="primary" type="submit">
-                        Save Changes
-                      </Button>
-                    </Modal.Footer>
-                  </Modal>
+                  <TaskModal singleTask={singleTask} />
                 </Card.Body>
               </Card>
             );
